+++ Assets/pocketbase-unity/Runtime/Services/HealthService.cs
@@ -16,15 +16,17 @@ namespace PocketBaseSdk
         /// Usually shouldn't be initialized manually and instead
         /// <cref cref="PocketBase.Health"/> should be used.
         /// </remarks>
-        public Task<HealthCheck> Check(
+        public async Task<HealthCheck> Check(
             Dictionary<string, object> query = null,
             Dictionary<string, string> headers = null)
         {
-            return _client.Send(
+            var result = await _client.Send(
                 "/api/health",
                 query: query,
                 headers: headers
-            ).ContinueWith(t => t.Result.ToObject<HealthCheck>());
+            );
+
+            return result.ToObject<HealthCheck>();
         }
     }
 }