+++ CHANGELOG.md
@@ -5,6 +5,58 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.23.5] - 2025-04-29
+
+### Removed
+
+- Task extension methods `ContinueWithOnMainThread` are no longer part of the SDK.
+
+## [0.23.4] - 2025-04-28
+
+### Fixed
+
+- Fixed task continuation not running on WebGL.
+
+### Changed
+
+- Await all tasks instead of using `ContinueWith()`, as this is not supported in Unity WebGL.
+
+## [0.23.3] - 2025-03-16
+
+### Deprecated
+
+- `RecordModel.Created` and `RecordModel.Updated` are now deprecated in favor of `RecordModel.Get<string>("created")` and `RecordModel.Get<string>("updated")` to be more consistent with the Dart SDK.
+
+### Added
+
+- `RecordModel.Get<T>(string, T)` method to extract a single value from the `RecordModel` by a dot-notation path and try to cast it to the specified generic type.
+- `RecordModel.GetListValue<T>(string, List<T>)` alias for `Get<List<T>>(string, List<T>)`.
+- `RecordModel.GetStringValue(string, string)` alias for `Get<string>(string, string)`.
+- `RecordModel.GetBoolValue(string, bool)` alias for `Get<bool>(string, bool)`.
+- `RecordModel.GetIntValue(string, int)` alias for `Get<int>(string, int)`.
+- `RecordModel.GetFloatValue(string, float)` alias for `Get<float>(string, float)`.
+
+### Changed
+
+- `RecordModel` design to be more consistent with the Dart SDK.
+- `RecordModel.Data` is now a `JObject` instead of a `Dictionary<string, JToken>`.
+
+## [0.23.2] - 2025-03-07
+
+### Changed
+
+- `RecordAuth.Record` is now of type `RecordModel` instead of `BaseAuthModel`. You can still access the user data via the `RecordModel[string]` indexer: `auth.Record["name"]`.
+
+### Removed
+
+- Removed `BaseAuthModel`.
+
+## [0.23.1] - 2025-02-06
+
+### Fixed
+
+- Replaced incorrect using directives (Unity.Plastic.Newtonsoft.Json and Codice.Utils) with the correct ones.
+
 ## [0.23.0] - 2025-01-30
 
 ### Changed